+++ src/cli/authn/authn-describe.ts
@@ -1,0 +1,41 @@
+import { frodo } from '@rockcarver/frodo-lib';
+import { Option } from 'commander';
+
+import { describeAuthenticationSettings } from '../../ops/AuthenticationSettingsOps';
+import { verboseMessage } from '../../utils/Console';
+import { FrodoCommand } from '../FrodoCommand';
+
+const { getTokens } = frodo.login;
+
+const program = new FrodoCommand('frodo authn describe');
+
+program
+  .description('Describe authentication settings.')
+  .addOption(new Option('--json', 'Output in JSON format.'))
+  .action(
+    // implement command logic inside action handler
+    async (host, realm, user, password, options, command) => {
+      command.handleDefaultArgsAndOpts(
+        host,
+        realm,
+        user,
+        password,
+        options,
+        command
+      );
+      if (await getTokens()) {
+        verboseMessage(`Describing authentication settings...`);
+        const outcome = await describeAuthenticationSettings(options.json);
+        if (!outcome) process.exitCode = 1;
+      }
+      // unrecognized combination of options or no options
+      else {
+        verboseMessage('Unrecognized combination of options or no options...');
+        program.help();
+        process.exitCode = 1;
+      }
+    }
+    // end command logic inside action handler
+  );
+
+program.parse();